+++ terraform/vpc-peering/requester/main.tf
@@ -14,8 +14,9 @@ module "core" {
   subnet                    = var.config.accepter_subnet
   vpc_peering_connection_id = aws_vpc_peering_connection.this.id
   security_group_map        = coalesce(var.config.security_group_map, {})
+  ecs_security_groups       = coalesce(var.config.ecs_security_groups, {})
   vpc_name                  = var.config.accepter_vpc_name
   source_vpc_id             = var.config.source_vpc_id
-  target_hosted_zone_id     = var.config.target_hosted_zone_id
+  target_hosted_zone_ids    = var.config.target_hosted_zone_ids
   accept_remote_dns         = var.config.accept_remote_dns
 }