+++ terraform/vpc-peering/requester/variables.tf
@@ -6,15 +6,20 @@ variable "name" {
 
 variable "config" {
   type = object({
-    accepter_account_id   = string
-    accepter_vpc          = string
-    accepter_vpc_name     = string
-    accepter_region       = optional(string)
-    accepter_subnet       = string
-    requester_vpc         = string
-    security_group_map    = optional(map(string))
-    source_vpc_id         = optional(string)
-    target_hosted_zone_id = optional(string)
-    accept_remote_dns     = optional(bool)
+    accepter_account_id    = string
+    accepter_vpc           = string
+    accepter_vpc_name      = string
+    accepter_region        = optional(string)
+    accepter_subnet        = string
+    requester_vpc          = string
+    source_vpc_id          = optional(string)
+    target_hosted_zone_ids = optional(list(string))
+    accept_remote_dns      = bool
+    security_group_map     = optional(map(string))
+    ecs_security_groups = optional(map(object({
+      port        = string
+      application = string
+      environment = string
+    })))
   })
 }